refactor: post/public으로 컴포넌트 & 서비스 이동
