+++ src/features/post/public/components/PostInformation.tsx
