+++ src/features/post/public/components/PostTitle.tsx
